requirements.txt
Django==2.2.3
django-bleach==0.5.3
django-summernote==0.8.11.4
bs4==0.0.1
gtts==2.0.3